requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
lxml==3.8.0
MarkupSafe==1.0
python-dateutil==2.6.1
requests==2.18.4
selenium==3.5.0
six==1.10.0
urllib3==1.22
Werkzeug==0.12.2